feat: clean up preset store
