+++ apps/codeimage/src/state/presets/types.ts
@@ -1,0 +1,4 @@
+import {GetPresetByIdApi} from '@codeimage/api/api-types';
+
+export type PresetsArray = Array<GetPresetByIdApi['response']>;
+export type Preset = GetPresetByIdApi['response'];
